requirements.txt
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Login==0.6.3
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
PyJWT==2.12.1
python-dotenv==1.2.2
typing_extensions==4.15.0
Werkzeug==3.1.7
wrapt==2.1.2
psycopg2-binary